+++ src/applications/static-pages/facilities/vetCentersHours.jsx
@@ -1,31 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { formatHours } from '../../facility-locator/utils/formatHours';
 
-// TODO fix spacing and add unit tests
 const VetCenterHours = props => {
   const arrayOfWeekdays = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
 
-  const formatTime = hoursString => {
-    // starthours: 700, endhours: 1730
-    let hour = Number(
-      hoursString.toString().length > 3
-        ? hoursString.toString().substring(0, 2)
-        : hoursString.toString().substring(0, 1),
-    );
-    const minutes = Number(
-      hoursString.toString().length > 3
-        ? hoursString.toString().substring(2, 4)
-        : hoursString.toString().substring(1, 4),
-    );
-    let AMorPM = 'a.m.';
-    if (hour >= 0 && hour <= 24 && minutes >= 0 && minutes <= 60) {
-      if (hour > 12) AMorPM = 'p.m.';
-      hour %= 12;
-    }
-    return `${hour}:${
-      minutes === 0 ? `${minutes}${minutes}` : `${minutes}`
-    } ${AMorPM}`;
-  };
   const buildHours = hours => {
     const hoursListItems = [...hours.slice(-6), hours[0]].map(hourObj => {
       // {day: 4, starthours: 700, endhours: 1730, comment: ""}
@@ -33,9 +12,9 @@ const VetCenterHours = props => {
       if (hourObj.starthours < 0 || hourObj.endhours < 0) {
         hourLabel = `${arrayOfWeekdays[hourObj.day]}: ${hourObj.comment}`;
       } else {
-        hourLabel = `${arrayOfWeekdays[hourObj.day]}: ${formatTime(
+        hourLabel = `${arrayOfWeekdays[hourObj.day]}: ${formatHours(
           hourObj.starthours,
-        )} - ${formatTime(hourObj.endhours)}`;
+        )} - ${formatHours(hourObj.endhours)}`;
       }
       return (
         <li className="vads-u-margin-bottom--0" key={hourObj.day}>
@@ -56,14 +35,14 @@ const VetCenterHours = props => {
         Hours
       </h3>
       <div className="vads-u-display--flex vads-u-flex-direction--column small-screen:vads-u-flex-direction--row vads-u-margin-bottom--0">
-        {buildHours(props.hoursMap)}
+        {buildHours(props.hours)}
       </div>
     </div>
   );
 };
 
 VetCenterHours.propTypes = {
-  hoursMap: PropTypes.array,
+  hours: PropTypes.array,
 };
 
 export default VetCenterHours;
